+++ src/util/serializing/ObjectOutputStream.cpp
@@ -46,7 +46,7 @@ void ObjectOutputStream::writeString(const char* str) { writeString(std::string(
 
 void ObjectOutputStream::writeString(const std::string& s) {
     this->encoder->addStr("_s");
-    int len = s.length();
+    int len = static_cast<int>(s.length());
     this->encoder->addData(&len, sizeof(int));
     this->encoder->addData(s.c_str(), len);
 }
@@ -69,7 +69,7 @@ void ObjectOutputStream::writeData(const std::vector<T>& data) {
     const int width = sizeof(T);
     this->encoder->addData(&len, sizeof(int));
     this->encoder->addData(&width, sizeof(int));
-    this->encoder->addData(data.data(), data.size() * sizeof(T));
+    this->encoder->addData(data.data(), static_cast<int>(data.size() * sizeof(T)));
 }
 
 void ObjectOutputStream::writeImage(const std::string_view& imgData) {
